Move formatBytecode to utils and rename to parseBytecode
